pyproject.toml
[tool.poetry]
name = "pythonproject5"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.15"
python-dotenv = "^1.0.0"
asyncpg = "^0.27.0"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"
fastapi-users = "^11.0.0"
uuid = "^1.30"
hasher = "^2.0.1"
python-jose = "^3.3.0"
greenlet = "^2.0.2"
httpx = "^0.24.1"
bs4 = "^0.0.1"
lxml = "^4.9.2"
pillow = "^9.5.0"
selenium = "^4.10.0"
webdriver-manager = "^3.8.6"
gunicorn = "^20.1.0"
fastapi = "^0.100.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
